// client/src/app/shared/model/profile-model.component.ts
export interface Profile {
  emailAddress: string;
  accountNumber: string;
  fullName: string;
  lastName: string;
  firstName: string;
  middleName: string;

  age: number;
  gender: string;
  dateOfBirth: string;
  address1: string;
  address2: string;
  city: string;
  province: string;
  contactNumber: string;
}